pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mathe-core-mlib"  # Nome de instalação (pip install mathe-core-mlib)
version = "0.1.0"
description = "Biblioteca base de utilitários matemáticos e gráficos."
authors = [{name = "Matheus Lôbo dos Santos", email = "matheus.lobo@ufpe.br"}]
dependencies = [
    "numpy",
    "scipy",
    "pyyaml",
    "pandas"
]

[tool.setuptools.packages.find]
where = ["src"]